requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.3.1=pyhd8ed1ab_0
astroid=2.4.2=py39hecd8cb5_0
autopep8=1.5.4=py_0
ca-certificates=2020.12.5=h033912b_0
certifi=2020.12.5=py39h6e9494a_1
django=3.1.5=pyhd8ed1ab_0
django-filter=2.4.0=py_0
djangorestframework=3.12.2=pyhd8ed1ab_0
djangorestframework-filters=1.0.0.dev0=py_0
isort=5.6.4=py_0
lazy-object-proxy=1.4.3=py39h9ed2024_2
libcxx=11.0.1=habf9029_0
libffi=3.3=h046ec9c_2
mccabe=0.6.1=py39hecd8cb5_1
ncurses=6.2=h2e338ed_4
openssl=1.1.1i=h35c211d_0
pip=21.0=pyhd8ed1ab_0
pycodestyle=2.6.0=py_0
pylint=2.6.0=py39hecd8cb5_0
pylint-django=2.4.2=pypi_0
pylint-plugin-utils=0.6=pypi_0
python=3.9.1=h1d169a7_3_cpython
python_abi=3.9=1_cp39
pytz=2020.5=pyhd8ed1ab_0
readline=8.0=h0678c8f_2
setuptools=49.6.0=py39h6e9494a_3
six=1.15.0=py39hecd8cb5_0
sqlite=3.34.0=h17101e1_0
sqlparse=0.4.1=pyh9f0ad1d_0
tk=8.6.10=h0419947_1
toml=0.10.1=py_0
tzdata=2020f=he74cb21_0
wheel=0.36.2=pyhd3deb0d_0
wrapt=1.11.2=py39h9ed2024_0
xz=5.2.5=haf1e3a3_1
zlib=1.2.11=h7795811_1010
